+++ common/pri_queue.h
@@ -1,7 +1,7 @@
 #ifndef ____pri_queue__
 #define ____pri_queue__
 
-typedef struct { void * data; int pri; } q_elem_t;
+typedef struct { void * data; long pri; } q_elem_t;
 typedef struct { q_elem_t *buf; int n, alloc; } pri_queue_t, *pri_queue;
 
 #endif /* defined(____pri_queue__) */
@@ -9,10 +9,10 @@ typedef struct { q_elem_t *buf; int n, alloc; } pri_queue_t, *pri_queue;
 
 pri_queue priq_new(int size);
 
-void priq_push(pri_queue q, void *data, int pri);
+void priq_push(pri_queue q, void *data, long pri);
 
-void *priq_pop(pri_queue q, int *pri);
+void *priq_pop(pri_queue q, long *pri);
 
-void *priq_top(pri_queue q, int *pri);
+void *priq_top(pri_queue q, long *pri);
  
 void priq_combine(pri_queue q, pri_queue q2);
